fix: Integrate duplicated features
