+++ client/client.h
@@ -1,15 +1,13 @@
 #ifndef CLIENT_H
+#define CLIENT_H
+
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
 #include <unistd.h>
 #include <arpa/inet.h>
 #include <sys/socket.h>
-
-#define CLIENT_H
-
-#define APP_NAME "TTT"
-#define APP_VERSION "v0.0.1"
+#include "../common/type.h"
 
 void connect_to_server();
