+++ client/util.c
@@ -5,42 +5,20 @@
 #include <fcntl.h>
 #include <curses.h>
 #include <string.h>
+#include <sys/ioctl.h>
 
-#define NUMBER_OF_CHAR_AT_A_TIME 1
 #define TITLE_START_LINE 3
 
-void tty_mode(int how) {
-	static struct termios orig_mode;
+int window_width = 0;
+int window_height = 0;
 
-	if (how == 0) 					// Save current setting
-		tcgetattr(0, &orig_mode); 
-	else if (how == 1)				// Recover saved setting
-		tcsetattr(0, TCSANOW, &orig_mode);
-}
+void get_window_size() {
+	struct winsize wbuf;
 
-void set_cr_noecho_mode() {
-	struct termios ttyinfo;
-	
-	if (tcgetattr(0, &ttyinfo) == -1) {
-		perror("tcgetattr");
-		exit(EXIT_FAILURE);
+	if (ioctl(0, TIOCGWINSZ, &wbuf) != -1) {
+		window_width = wbuf.ws_col;
+		window_height = wbuf.ws_row;
 	}
-
-	ttyinfo.c_lflag &= ~ICANON;		// Set non-canonical mode
-	ttyinfo.c_lflag &= ~ECHO;		// Disable ECHO bit
-	ttyinfo.c_cc[VMIN] = NUMBER_OF_CHAR_AT_A_TIME;
-
-	if (tcsetattr(0, TCSANOW, &ttyinfo) == -1) {
-		perror("tcsetattr");
-		exit(EXIT_FAILURE);
-	}
-}
-
-void set_non_blocking_mode() {
-    int terflags;
-    terflags = fcntl(0, F_GETFL);
-    terflags |= O_NONBLOCK;
-    fcntl(0, F_SETFL, terflags);
 }
 
 void show_title(const char* title) {
@@ -53,3 +31,4 @@ void show_title(const char* title) {
         printw("=");
 	printw("\n\n");
 }
+
